requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.26.0
anthropic>=0.25.0
python-docx>=1.1.0
pdfplumber>=0.10.0
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
aiohttp>=3.9.0
requests>=2.31.0
